requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
python-dotenv==1.0.1
requests==2.32.3
pytest==7.4.2 #added
pytest-mock==3.11.1 #added
bcrypt==4.0.1 #added